Update example and fix smaller issues
