+++ example/test/example_test.dart
@@ -1,8 +1,7 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  test('test', () {
+  test('Example test', () {
     //
   });
 }
