--- res/values/strings_en.arb
@@ -1,1 +1,0 @@
-{}
